+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# callr development version
+
 # callr 3.5.1
 
 * `callr::r_session` now handles large messages from the subprocess